test cases removed, will be brought back
